google_data/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
gspread
google-auth
requests
httpx
openai
pypdf>=4.2.0
python-docx>=0.8.11
